requirements.txt
pandas>=2.1,<3
geopandas>=1.0,<2
pyarrow>=14,<24
PyYAML>=6.0,<7
requests>=2.31,<3
streamlit>=1.35,<2
plotly>=5.18,<7
folium>=0.15,<1
streamlit-folium>=0.18,<1
python-pptx>=1.0,<2
